Extract intent utils to separate file
